requirements.txt
torch==2.6.0+cu118
torchvision==0.21.0+cu118
numpy>=2.2.4
pandas>=2.2.3
pillow>=11.1.0
scikit-learn>=1.6.1
torch_geometric>=2.6.1